Restrict and Require Attribute@Feature
The attribute Feature in the node Attribute is implicitly required
by DIN 15800:2020-07 on page 12, because the attribute ActivationGroup
is explicitly optional. Also, the format was restricted to two
nametypes, separated by a dot.

Unfortunately, XSD 1.0 can't validate that Feature points to a valid
Feature, because xs:keyref is not expressive enough to split a string
and check that it resolves to a valid node.
This might be solvable with xs:assert in XSD 1.1.