Remove RDM Substitution for FTRDM again
The name RDM for the name is officially confirmed to be a typo. FTRDM
is the official name of the node, see
https://gdtf-share.com/forum/index.php?/topic/444-rdm-node-has-inconsistent-name-in-din-standard/

While it is my interpretation that a file using the name RDM would
be a valid GDTF according to the DIN 15800:2020-07, it is probably
bad practice to encourage usage of it by allowing it in this schema.
Better to force everyone to use the pseudo-standard established by
the GDTF builder and GrandMA 3.
Also, if this schema is more strict than the standard itself, it should
not pose many problems. It would be more dangerous if the schema were
more lax than the standard.