Allow Arbitrary Order of Properties Children
This means that in XSD 1.0 the occurence cannot be properly
constrained. But otherwise some files fail even though they shouldn't.

An ugly workaround would be to write out every one of the 24 possible
orders and wrap the sequences in a single choice. I think the better
alternative is to add a second schema language, like Schematron, where
the occurence can be validated.
This is needed to validate the nested (dot-separated) references
anyway.